requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
easyocr
exceptiongroup
faiss-cpu
filelock
frozenlist
fsspec
gitdb
GitPython
huggingface-hub
idna
imageio
importlib-metadata
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-experimental
langsmith
lazy_loader
llvmlite
markdown-it-py
MarkupSafe
marshmallow
mdurl
more-itertools
mpmath
multidict
mypy-extensions
mysql-connector-python
networkx
ninja
numba
numpy
openai
openai-whisper
opencv-python-headless
packaging
pandas
Pillow
protobuf
pyarrow
pyclipper
pydantic
pydantic_core
pydeck
Pygments
PyMySQL
PyPDF2
python-bidi
python-dateutil
python-dotenv
python-http-client
pytube
pytz
PyYAML
referencing
regex
requests
rich
rpds-py
scikit-image
scipy
semantic-version
sendgrid
setuptools-rust
shapely
six
smmap
sniffio
SQLAlchemy
starkbank-ecdsa
streamlit
sympy
tenacity
tifffile
tiktoken
tokenizers
toml
tomli
toolz
torch
torchvision
tornado
tqdm
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
validators
yarl
zipp
